password reset

+++ src/main/java/pl/hipotrofia/controllers/UserController.java
@@ -1,5 +1,7 @@
 package pl.hipotrofia.controllers;
 
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import pl.hipotrofia.converters.UserDtoConverter;
 import pl.hipotrofia.dto.UserDto;
@@ -7,11 +9,11 @@ import pl.hipotrofia.entities.User;
 import pl.hipotrofia.services.RoleService;
 import pl.hipotrofia.services.UserService;
 
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 @RestController
 @RequestMapping("api/users")
-//@PreAuthorize("hasRole('ROLE_ADMIN')")
 public class UserController {
 
     private final UserService userService;
@@ -23,43 +25,85 @@ public class UserController {
                           RoleService roleService) {
         this.userDtoConverter = userDtoConverter;
         this.userService = userService;
-        this.roleService=roleService;
+        this.roleService = roleService;
     }
 
-    @GetMapping("/anonymous/getAll/{limit}/{offset}")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @GetMapping("/getAll/{limit}/{offset}")
     public List<UserDto> getAll(@PathVariable int limit, @PathVariable int offset) {
 
         return userDtoConverter.convertToDto(userService.getAllLimited(limit, offset));
     }
 
+    @PreAuthorize("hasAnyRole('ROLE_USER', 'ROLE_ADMIN', 'ROLE_PUBLISHER')")
     @DeleteMapping("/delete")
-    public boolean delete(@RequestParam Long id) {
+    public void delete(@RequestParam Long id, HttpServletResponse response) {
+
+        final String email = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        final String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
 
         try {
             User user = userService.findUserById(id);
-            userService.delete(user);
-            return true;
+            if (role.equals("[ROLE_ADMIN]") ||  user.getEmail().equals(email)) {
+                userService.delete(user);
+            } else {
+                response.setStatus(405);
+                response.setHeader("ERROR", "Something went wrong...");
+            }
         } catch (Exception ex) {
+            response.setHeader("ERROR", ex.getMessage());
             ex.printStackTrace();
-            return false;
         }
     }
 
+    @PreAuthorize("hasAnyRole('ROLE_USER', 'ROLE_ADMIN', 'ROLE_PUBLISHER')")
     @PostMapping("/edit")
-    public boolean edit(@RequestBody UserDto userDto) {
+    public void edit(@RequestBody UserDto userDto, HttpServletResponse response) {
+
+        final String email = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+
+        if (email.equals(userDto.getEmail())) {
+            try {
+                User user = userService.findUserById(userDto.getId());
+                user.setActive(userDto.isActive());
+                user.setRole(roleService.findByName(userDto.getRoleName()));
+                user.setName(userDto.getName());
+
+                userService.save(user);
+                response.setHeader("EDIT", "Success");
+            } catch (Exception ex) {
+                response.setHeader("ERROR", ex.getMessage());
+                ex.printStackTrace();
+            }
+        } else {
+            response.setStatus(405);
+            response.setHeader("ERROR", "Something went wrong...");
+        }
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PostMapping("/deactivate")
+    public void deactivate(@RequestBody UserDto userDto, HttpServletResponse response) {
 
         try {
-            User user = userService.findUserById(userDto.getId());
-            user.setActive(userDto.isActive());
-            user.setRole(roleService.findByName(userDto.getRoleName()));
-            user.setName(userDto.getName());
-
-            userService.save(user);
-
-            return true;
+            User user = userService.findUserByEmail(userDto.getEmail());
+            userService.deactivate(user);
         } catch (Exception ex) {
+            response.setHeader("ERROR", ex.getMessage());
             ex.printStackTrace();
-            return false;
+        }
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PostMapping("/activate")
+    public void setActive(@RequestBody UserDto userDto, HttpServletResponse response) {
+
+        try {
+            User user = userService.findUserByEmail(userDto.getEmail());
+            userService.activate(user);
+        } catch (Exception ex) {
+            response.setHeader("ERROR", ex.getMessage());
+            ex.printStackTrace();
         }
     }
